sdm845-common: DeviceSettings: Enable dither with DC dimming
DC dimming (LineageOS `Anti Flicker`) works by overlaying a progressively more-opaque layer of black over the rest of the image as brightness decreases, which effectively reduces the number of colors between 0 and fully-on for each color channel (R, G, B).
It's the equivalent of setting your computer display to 16- or 8-bit color mode; you get ugly banding and colors are "off" or crushed when they're near-white or near-black.
Enabling dither helps to perceptually improve the image by adding a noise pattern to boundaries between available colors to remove the appearance of banding/crush.

This ties the 'dither_en' kernel node state (if it exists) to 'dimlayer_bl_en' state in sysfs (if it exists) to enable dither whenever DC dimming is toggled.

Squashing in three commits from a closed PR from a deleted branch, since I decided to go with DeviceSettings instead of LiveDisplay after all,
and Leah is awesome and deserves all the credit for making this work:

* add `dither_en`

* Update DCModeSwitch.java

* make it work™